+++ packages/expo-server/src/runtime/index.ts
@@ -1,5 +1,5 @@
 import { errorToResponse } from './error';
-import { type ScopeDefinition, type RequestAPI, scopeRef } from './scope';
+import { type ScopeDefinition, type RequestAPI, scopeRef, UpdateResponseHeaders } from './scope';
 import { importMetaRegistry } from '../utils/importMetaRegistry';
 
 export interface RequestAPISetup extends RequestAPI {
@@ -56,6 +56,8 @@ export function createRequestScope<F extends RequestContextFactory>(
 
     const setup = makeRequestAPISetup(...args);
     const { waitUntil = defaultWaitUntil } = setup;
+    const deferredTasks: (() => Promise<unknown> | void)[] = [];
+    const responseHeadersUpdates: UpdateResponseHeaders[] = [];
 
     const scope = {
       ...setup,
@@ -63,9 +65,11 @@ export function createRequestScope<F extends RequestContextFactory>(
       environment: setup.environment,
       waitUntil,
       deferTask: setup.deferTask,
+      setResponseHeaders(updateHeaders) {
+        responseHeadersUpdates.push(updateHeaders);
+      },
     } satisfies RequestAPI;
 
-    const deferredTasks: (() => Promise<unknown>)[] = [];
     if (!scope.deferTask) {
       scope.deferTask = function deferTask(fn) {
         deferredTasks.push(fn);
@@ -79,15 +83,42 @@ export function createRequestScope<F extends RequestContextFactory>(
           ? await scopeRef.current.run(scope, () => run(...args))
           : await run(...args);
     } catch (error) {
-      if (error != null && error instanceof Error && 'status' in error) {
+      if (error != null && error instanceof Response && !error.bodyUsed) {
+        result = error;
+      } else if (error != null && error instanceof Error && 'status' in error) {
         return errorToResponse(error);
       } else {
         throw error;
       }
     }
 
-    if (deferredTasks.length) {
-      deferredTasks.forEach((fn) => waitUntil(fn()));
+    deferredTasks.forEach((fn) => {
+      const maybePromise = fn();
+      if (maybePromise != null) waitUntil(maybePromise);
+    });
+
+    for (const updateHeaders of responseHeadersUpdates) {
+      let headers: Headers = result.headers;
+      if (typeof updateHeaders === 'function') {
+        headers = updateHeaders(result.headers) || headers;
+      } else if (updateHeaders instanceof Headers) {
+        headers = updateHeaders;
+      } else if (typeof updateHeaders === 'object' && updateHeaders) {
+        for (const headerName in updateHeaders) {
+          if (Array.isArray(updateHeaders[headerName])) {
+            for (const headerValue of updateHeaders[headerName]) {
+              headers.append(headerName, headerValue);
+            }
+          } else if (updateHeaders[headerName] != null) {
+            headers.set(headerName, updateHeaders[headerName]);
+          }
+        }
+      }
+      if (headers !== result.headers) {
+        for (const [headerName, headerValue] of headers) {
+          result.headers.set(headerName, headerValue);
+        }
+      }
     }
 
     return result;